% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDinterfaces.R
\name{print.FDfitted}
\alias{print.FDfitted}
\title{Print Method for FDfitted Objects}
\usage{
\method{print}{FDfitted}(x, ...)
}
\arguments{
\item{x}{an object of class FDfitted, usually the result of \code{\link{FD.estimation}}.}

\item{...}{additional arguments}
}
\description{
This method shows the results of \code{\link{FD.estimation}}.
}
